# admin-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "i18n-flow 是一个用于管理多语言翻译的系统。",
        "title": "i18n-flow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/exports/project/{project_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "导出指定项目的所有翻译，支持不同格式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "导出项目翻译",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "项目ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "导出格式，默认json",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/imports/project/{project_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为指定项目导入翻译数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "导入项目翻译",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "项目ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "翻译数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/translations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为特定项目和语言创建新的翻译条目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "创建翻译条目",
                "parameters": [
                    {
                        "description": "翻译信息",
                        "name": "translation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.TranslationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/service.TranslationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translations/batch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为特定项目批量创建多种语言的翻译",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "批量创建翻译",
                "parameters": [
                    {
                        "description": "批量翻译信息，translations 字段为语言代码到翻译值的映射，如 {'zh-CN':'中文值','en':'English value'}",
                        "name": "translations",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.BatchTranslationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.TranslationResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translations/batch-delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "批量删除多个翻译条目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "批量删除翻译",
                "parameters": [
                    {
                        "description": "翻译ID数组",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translations/by-project/{project_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定项目的所有翻译条目，支持分页和搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "获取项目翻译",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "项目ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "通过ID获取特定翻译条目的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "获取翻译详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "翻译ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.TranslationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新特定翻译条目的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "更新翻译",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "翻译ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "翻译信息",
                        "name": "translation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.TranslationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.TranslationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除特定的翻译条目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "翻译管理"
                ],
                "summary": "删除翻译",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "翻译ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.BatchTranslationRequest": {
            "type": "object",
            "required": [
                "key_name",
                "project_id"
            ],
            "properties": {
                "context": {
                    "description": "上下文说明",
                    "type": "string",
                    "example": "欢迎消息"
                },
                "key_name": {
                    "description": "翻译键名",
                    "type": "string",
                    "example": "welcome_message"
                },
                "project_id": {
                    "description": "项目ID",
                    "type": "integer",
                    "example": 1
                },
                "translations": {
                    "description": "语言代码 -\u003e 翻译值",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "'en'": "'Welcome'}",
                        "{'zh-CN'": "'欢迎'"
                    }
                }
            }
        },
...
        "service.TranslationRequest": {
            "type": "object",
            "required": [
                "key_name",
                "language_id",
                "project_id"
            ],
            "properties": {
                "context": {
                    "description": "上下文说明",
                    "type": "string",
                    "example": "主页欢迎消息"
                },
                "key_name": {
                    "description": "翻译键名",
                    "type": "string",
                    "example": "welcome_message"
                },
                "language_id": {
                    "description": "语言ID",
                    "type": "integer",
                    "example": 2
                },
                "project_id": {
                    "description": "项目ID",
                    "type": "integer",
                    "example": 1
                },
                "value": {
                    "description": "翻译值",
                    "type": "string",
                    "example": "欢迎使用"
                }
            }
        },
        "service.TranslationResponse": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "上下文说明",
                    "type": "string",
                    "example": "欢迎消息"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2023-04-01 12:00:00"
                },
                "id": {
                    "description": "翻译ID",
                    "type": "integer",
                    "example": 1
                },
                "key_name": {
                    "description": "翻译键名",
                    "type": "string",
                    "example": "welcome_message"
                },
                "language_code": {
                    "description": "语言代码",
                    "type": "string",
                    "example": "zh-CN"
                },
                "language_id": {
                    "description": "语言ID",
                    "type": "integer",
                    "example": 2
                },
                "language_name": {
                    "description": "语言名称",
                    "type": "string",
                    "example": "简体中文"
                },
                "project_id": {
                    "description": "项目ID",
                    "type": "integer",
                    "example": 1
                },
                "project_name": {
                    "description": "项目名称",
                    "type": "string",
                    "example": "网站翻译"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "active"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2023-04-01 12:00:00"
                },
                "value": {
                    "description": "翻译值",
                    "type": "string",
                    "example": "欢迎使用"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}